wildlife_poaching_detection/src/backend/requirements.txt
# Wildlife Protection AI System - Backend Dependencies
# Generated for Production Deployment - November 8, 2025

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning - PyTorch (Apple Silicon optimized)
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# Machine Learning - TensorFlow (Apple Silicon optimized) 
tensorflow-macos==2.13.0
tensorflow-metal==1.0.1

# Computer Vision & Image Processing
opencv-python==4.8.1.78
Pillow==10.0.1
ultralytics==8.0.206

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# Data Processing
numpy==1.24.3
pandas==2.1.3

# Utilities
python-dateutil==2.8.2
pathlib2==2.3.7.post1
requests==2.31.0

# Logging & Monitoring
loguru==0.7.2

# File Handling
aiofiles==23.2.1

# Development (Optional)
pytest==7.4.3
httpx==0.25.2

# Note: Install with Apple Silicon optimization
# pip install -r requirements.txt
# 
# For Apple Silicon users, ensure you have:
# - tf-macos virtual environment activated
# - Xcode Command Line Tools installed
# - conda install pytorch torchvision torchaudio -c pytorch (if using conda)